# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local Development"
    },
    {
      "url": "https://staging-api.yourdomain.com/api/v1",
      "description": "Staging Environment"
    },
    {
      "url": "https://api.yourdomain.com/api/v1",
      "description": "Production Environment"
    }
  ],
  "info": {
    "title": "Omni-Backend API",
    "version": "1.0.0",
    "description": "Generated by nestia - https://github.com/samchon/nestia",
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "schemas": {
      "PlatformAccess": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "X",
              "Instagram",
              "LinkedIn",
              "Telegram",
              "Email"
            ]
          },
          "canSend": {
            "type": "boolean"
          },
          "viewMessages": {
            "type": "boolean"
          }
        },
        "required": [
          "platform",
          "canSend",
          "viewMessages"
        ]
      },
      "RoleEntity": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "PA",
              "super-admin"
            ]
          },
          "user": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityDTOTT.o2"
            }
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "__@OptionalProps@383": {
            "type": "string",
            "enum": [
              "createdAt",
              "updatedAt"
            ]
          }
        },
        "required": [
          "name",
          "user",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "EntityDTOTT.o2": {
        "type": "object",
        "properties": {},
        "required": []
      },
      "RoleEntity__loadedTypeRoleEntityundefined": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "PA",
              "super-admin"
            ]
          },
          "user": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityDTOTT.o2"
            }
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "__@OptionalProps@383": {
            "type": "string",
            "enum": [
              "createdAt",
              "updatedAt"
            ]
          },
          "__@___loadedType@1012": {
            "$ref": "#/components/schemas/RoleEntity"
          }
        },
        "required": [
          "name",
          "user",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  },
  "paths": {
...
    "/user/all": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": false
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "firstName": {
                            "type": "string"
                          },
                          "lastName": {
                            "type": "string"
                          },
                          "phoneNumber": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "inactive",
                              "disabled",
                              "pending"
                            ]
                          },
                          "disabled": {
                            "type": "boolean"
                          },
                          "emailVerifiedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "platformAccess": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/PlatformAccess"
                            }
                          },
                          "role": {
                            "$ref": "#/components/schemas/RoleEntity__loadedTypeRoleEntityundefined"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "email",
                          "firstName",
                          "lastName",
                          "status",
                          "platformAccess",
                          "role",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "offset": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "users",
                    "total",
                    "limit",
                    "offset"
                  ]
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "user",
          "all",
          "getAllUsers"
        ],
        "x-samchon-controller": "UserController"
      }
    },
    "/user/disable/{targetUserId}": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          }
        },
        "x-samchon-accessor": [
          "user",
          "disable",
          "disableUser"
        ],
        "x-samchon-controller": "UserController"
      }
    },
    "/user/enable/{targetUserId}": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          }
        },
        "x-samchon-accessor": [
          "user",
          "enable",
          "enableUser"
        ],
        "x-samchon-controller": "UserController"
      }
    },
    "/user/permissions/{targetUserId}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlatformAccess"
                  }
                }
              }
            }
          }
        },
        "x-samchon-accessor": [
          "user",
          "permissions",
          "viewPermissions"
        ],
        "x-samchon-controller": "UserController"
      },
      "patch": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PlatformAccess"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          }
        },
        "x-samchon-accessor": [
          "user",
          "permissions",
          "updateUserPlatformAccess"
        ],
        "x-samchon-controller": "UserController"
      }
    },
    "/user/{targetUserId}": {
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "targetUserId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          }
        },
        "x-samchon-accessor": [
          "user",
          "deleteUser"
        ],
        "x-samchon-controller": "UserController"
      }
    },
...
  },
...
}